+++ src/utils/decorators/should-component-update/__spec__.js
@@ -43,4 +43,35 @@ describe('Should Component Update Decorator', () => {
       expect(instance.shouldComponentUpdate(instance, nextProps, nextState)).toBeTruthy();
     });
   });
+
+  describe('displayName', () => {
+    class Foo extends React.Component { // eslint-disable-line react/no-multi-comp
+      bar = () => {
+        return 'bar';
+      }
+    }
+
+    const displayName = 'FooClass';
+
+    describe('when ComposedComponent.displayName is defined', () => {
+      beforeEach(() => {
+        Foo.displayName = displayName;
+      });
+      afterEach(() => {
+        Foo.displayName = undefined;
+      });
+
+      it('sets Component.displayName to ComposedComponent.displayName', () => {
+        const DecoratedComponent = ShouldComponentUpdateDecorator(Foo);
+        expect(DecoratedComponent.displayName).toBe(displayName);
+      });
+    });
+
+    describe('when ComposedComponent.displayName is undefined', () => {
+      it('sets Component.displayName to ComposedComponent.name', () => {
+        const DecoratedComponent = ShouldComponentUpdateDecorator(Foo);
+        expect(DecoratedComponent.displayName).toBe('Foo');
+      });
+    });
+  });
 });