+++ src/utils/decorators/tooltip-decorator/__spec__.js
@@ -623,4 +623,35 @@ describe('tooltip-decorator', () => {
       expect(noTooltip.getTooltip()).toEqual(null);
     });
   });
+
+  describe('displayName', () => {
+    class Foo extends React.Component { // eslint-disable-line react/no-multi-comp
+      bar = () => {
+        return 'bar';
+      }
+    }
+
+    const displayName = 'FooClass';
+
+    describe('when ComposedComponent.displayName is defined', () => {
+      beforeEach(() => {
+        Foo.displayName = displayName;
+      });
+      afterEach(() => {
+        Foo.displayName = undefined;
+      });
+
+      it('sets Component.displayName to ComposedComponent.displayName', () => {
+        const DecoratedComponent = TooltipDecorator(Foo);
+        expect(DecoratedComponent.displayName).toBe(displayName);
+      });
+    });
+
+    describe('when ComposedComponent.displayName is undefined', () => {
+      it('sets Component.displayName to ComposedComponent.name', () => {
+        const DecoratedComponent = TooltipDecorator(Foo);
+        expect(DecoratedComponent.displayName).toBe('Foo');
+      });
+    });
+  });
 });