+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Changed type of prop `tooltipMessage` of `tooltip-decorator` from string to node to allow children.
 
+## Fixes
+
+* Set the `Component.displayName` on all decorators. **NB** You may need to update your snapshots as a result of this to change the component name to something more accurate
+
 # 2.4.0
 
 ## Improvements